use size_t to align to typical cache line size, improve some error messages

+++ ocp_qp/x_ocp_qcqp.c
@@ -248,7 +248,7 @@ void OCP_QCQP_CREATE(struct OCP_QCQP_DIM *dim, struct OCP_QCQP *qp, void *mem)
 
 
 	// align to typical cache line size
-	long long l_ptr = (long long) i_ptr;
+	size_t l_ptr = (size_t) i_ptr;
 	l_ptr = (l_ptr+63)/64*64;
