+++ tree_ocp_qp/x_tree_ocp_qcqp_sol.c
@@ -131,7 +131,7 @@ void TREE_OCP_QCQP_SOL_CREATE(struct TREE_OCP_QCQP_DIM *dim, struct TREE_OCP_QCQ
 
 
 	// align to typical cache line size
-	long long l_ptr = (long long) sv_ptr;
+	size_t l_ptr = (size_t) sv_ptr;
 	l_ptr = (l_ptr+63)/64*64;
